REF: Rename files, vendor `unit_to_string`, update to new `openff-utilities`
